chore(setup): extract utilities into dedicated file
